+++ packages/rsc/src/browser.tsx
@@ -1,7 +1,7 @@
+import { memoize, tinyassert } from "@hiogawa/utils";
 import React from "react";
 import ReactDomClient from "react-dom/client";
 import ReactClient from "react-server-dom-webpack/client.browser";
-import { initializeReactClientBrowser } from "./features/client-component/browser";
 import type { RscPayload } from "./server";
 import type { CallServerCallback } from "./types";
 import { getRscScript } from "./utils/rsc-script";
@@ -131,4 +131,33 @@ function listenNavigation(onNavigation: () => void) {
     window.history.pushState = oldPushState;
     window.history.replaceState = oldReplaceState;
   };
+}
+
+//
+// client reference manifest
+//
+
+async function importClientRefrence(id: string) {
+  if (import.meta.env.DEV) {
+    // use raw import (inject in head script) to avoid `?import` added by vite import analysis
+    return (self as any).__raw_import(id);
+  } else {
+    const clientReferences = await import(
+      "virtual:client-references" as string
+    );
+    const dynImport = clientReferences.default[id];
+    tinyassert(dynImport, `client reference not found '${id}'`);
+    return dynImport();
+  }
+}
+
+let init = false;
+function initializeReactClientBrowser(): void {
+  if (init) return;
+  init = true;
+
+  Object.assign(globalThis, {
+    __webpack_require__: memoize(importClientRefrence),
+    __webpack_chunk_load__: async () => {},
+  });
 }
