--- packages/rsc/src/features/client-component/ssr.ts
@@ -1,50 +1,0 @@
-import { memoize, tinyassert } from "@hiogawa/utils";
-import type { ImportManifestEntry, ModuleMap } from "../../types";
-
-async function importClientReference(id: string) {
-  if (import.meta.env.DEV) {
-    return import(/* @vite-ignore */ id);
-  } else {
-    const clientReferences = await import(
-      "virtual:client-references" as string
-    );
-    const dynImport = clientReferences.default[id];
-    tinyassert(dynImport, `client reference not found '${id}'`);
-    return dynImport();
-  }
-}
-
-let init = false;
-export function initializeReactClientSsr(): void {
-  if (init) return;
-  init = true;
-
-  Object.assign(globalThis, {
-    __webpack_require__: memoize(importClientReference),
-    __webpack_chunk_load__: async () => {},
-  });
-}
-
-export function createModuleMap(): ModuleMap {
-  return new Proxy(
-    {},
-    {
-      get(_target, id, _receiver) {
-        return new Proxy(
-          {},
-          {
-            get(_target, name, _receiver) {
-              tinyassert(typeof id === "string");
-              tinyassert(typeof name === "string");
-              return {
-                id,
-                name,
-                chunks: [],
-              } satisfies ImportManifestEntry;
-            },
-          },
-        );
-      },
-    },
-  );
-}
