+++ packages/rsc/src/server.tsx
@@ -1,11 +1,7 @@
+import { memoize, tinyassert } from "@hiogawa/utils";
 import type { ReactFormState } from "react-dom/client";
 import ReactServer from "react-server-dom-webpack/server.edge";
-import { createBundlerConfig } from "./features/client-component/server";
-import {
-  createActionBundlerConfig,
-  importServerAction,
-  initializeReactServer,
-} from "./features/server-function/server";
+import type { BundlerConfig, ImportManifestEntry } from "./types";
 
 export type RscPayload = {
   root: React.ReactNode;
@@ -78,4 +74,84 @@ async function importSsrEntry(): Promise<typeof import("./ssr")> {
   } else {
     return await import("virtual:build-ssr-entry" as any);
   }
+}
+
+//
+// client reference manifest
+//
+
+function createBundlerConfig(): BundlerConfig {
+  return new Proxy(
+    {},
+    {
+      get(_target, $$id, _receiver) {
+        tinyassert(typeof $$id === "string");
+        let [id, name] = $$id.split("#");
+        tinyassert(id);
+        tinyassert(name);
+        return {
+          id,
+          name,
+          // TODO: preinit not working?
+          // `ReactDOMSharedInternals.d.X` seems no-op due to null request context?
+          chunks: [id, id],
+          async: true,
+        } satisfies ImportManifestEntry;
+      },
+    },
+  );
+}
+
+//
+// server reference manifest
+//
+
+async function importServerAction(id: string): Promise<Function> {
+  const [file, name] = id.split("#") as [string, string];
+  const mod: any = await importServerReference(file);
+  return mod[name];
+}
+
+async function importServerReference(id: string): Promise<unknown> {
+  if (import.meta.env.DEV) {
+    return import(/* @vite-ignore */ id);
+  } else {
+    const references = await import("virtual:server-references" as string);
+    const dynImport = references.default[id];
+    tinyassert(dynImport, `server reference not found '${id}'`);
+    return dynImport();
+  }
+}
+
+export function createActionBundlerConfig(): BundlerConfig {
+  return new Proxy(
+    {},
+    {
+      get(_target, $$id, _receiver) {
+        tinyassert(typeof $$id === "string");
+        let [id, name] = $$id.split("#");
+        tinyassert(id);
+        tinyassert(name);
+        return {
+          id,
+          name,
+          chunks: [],
+          async: true,
+        } satisfies ImportManifestEntry;
+      },
+    },
+  );
+}
+
+let init = false;
+function initializeReactServer(): void {
+  if (init) return;
+  init = true;
+
+  Object.assign(globalThis, {
+    __vite_react_server_webpack_require__: memoize(importServerReference),
+    __vite_react_server_webpack_chunk_load__: () => {
+      throw new Error("__webpack_chunk_load__");
+    },
+  });
 }
