+++ packages/rsc/src/ssr.tsx
@@ -1,10 +1,8 @@
+import { memoize, tinyassert } from "@hiogawa/utils";
 import ReactDomServer from "react-dom/server.edge";
 import ReactClient from "react-server-dom-webpack/client.edge";
-import {
-  createModuleMap,
-  initializeReactClientSsr,
-} from "./features/client-component/ssr";
 import type { RscPayload } from "./server";
+import type { ImportManifestEntry, ModuleMap } from "./types";
 import {
   createBufferedTransformStream,
   injectRscScript,
@@ -44,4 +42,56 @@ export async function renderHtml(stream: ReadableStream): Promise<Response> {
       "content-type": "text/html;charset=utf-8",
     },
   });
+}
+
+//
+// client reference manifest
+//
+
+async function importClientReference(id: string) {
+  if (import.meta.env.DEV) {
+    return import(/* @vite-ignore */ id);
+  } else {
+    const clientReferences = await import(
+      "virtual:client-references" as string
+    );
+    const dynImport = clientReferences.default[id];
+    tinyassert(dynImport, `client reference not found '${id}'`);
+    return dynImport();
+  }
+}
+
+let init = false;
+export function initializeReactClientSsr(): void {
+  if (init) return;
+  init = true;
+
+  Object.assign(globalThis, {
+    __webpack_require__: memoize(importClientReference),
+    __webpack_chunk_load__: async () => {},
+  });
+}
+
+export function createModuleMap(): ModuleMap {
+  return new Proxy(
+    {},
+    {
+      get(_target, id, _receiver) {
+        return new Proxy(
+          {},
+          {
+            get(_target, name, _receiver) {
+              tinyassert(typeof id === "string");
+              tinyassert(typeof name === "string");
+              return {
+                id,
+                name,
+                chunks: [],
+              } satisfies ImportManifestEntry;
+            },
+          },
+        );
+      },
+    },
+  );
 }
